chore: Setup prettier and eslint
Add basic rules that are in line with vue3, typescript, nuxt and tailwindcss.

Set allowed warnings to 0 as this is an anti-pattern and causes developers to ignore warnings.
Add workspace settings that allow eslint to fix where possible when saving and to
lint vue files properly.
Turn multi-template-root eslint rule off as multiple roots are allowed in vue3.

close: #9